A real Postgres in the tests, and the measurement that changed the policy
No Docker and no install: postgresql_embedded downloads and starts a
real Postgres per test. It found a bug in the first run -- attndims is
int2 in the catalog, not int4 -- which is precisely the class of error
that only a real database catches, because it is a wrong belief about
Postgres rather than a mistake in logic.

Six integration tests against it, including one that proves a refusal
was CORRECT by showing Postgres rejects the same insert. The oracle
works.

THEN THE MILESTONE, on schemas nobody here wrote:

  synapse   134 tables, 96% reach
  gitlab    955 tables,  8% reach

That split is the whole reason for measuring before generating. It is
not that GitLab is complex. Of its 2,537 CHECK expressions:

  1657  char_length(col) <= N       65%
   414  col IS NOT NULL             16%

81% are two shapes this ALREADY satisfies by construction. A length
limit is what varchar(n) means, and a NOT NULL written as a CHECK is a
column this never writes NULL into. Refusing them bought no safety at
all and cost 92% of the schema.

So checks.rs recognises a CLOSED SET of exact shapes and refuses
everything else. That is not the expression parser the plan rejected --
there is no evaluation, no simplification, no reasoning about
operators. A definition matches one of two literal forms or it is
Unknown. Nine shapes are tested to confirm they stay Unknown, including
`(file IS NULL) OR (char_length(file) <= 255)`, which exists to prove
the parenthesis stripper cannot be tricked into matching half of a
disjunction.

  gitlab   8% -> 52% reach, 878 refusals -> 458

Synapse unchanged at 96%.

GitLab still misses the 60% gate written down before any of this ran,
and that stands as a fail rather than being adjusted. The remaining 309
genuine CHECK refusals drag 282 more tables with them by contagion, and
the largest single shape left is octet_length(col) = N, 85 of them -- a
fixed-length bytea, which is trivially satisfiable and would be a third
literal shape. Whether to add it is a decision, not an automatic yes.

+++ src/introspect.rs
@@ -217,7 +217,11 @@ pub fn read(client: &mut Client, schemas: &[String]) -> Result<Schema, postgres:
             &typname,
             &(typtype as u8 as char).to_string(),
             row.get::<_, i32>("atttypmod"),
-            row.get::<_, i32>("attndims"),
+            // `attndims` is int2 in the catalog, not int4. Reading it as the
+            // wrong width is a deserialisation error rather than a silent
+            // misread, which is the good kind of wrong — and exactly the kind
+            // only a real database finds.
+            row.get::<_, i16>("attndims") as i32,
             base_typname.as_deref().zip(base_typtype.as_deref()),
             row.get::<_, bool>("domain_checked"),
             element_typname.as_deref().zip(element_typtype.as_deref()),
